+++ jms-ws-domain/src/main/java/com/jms/domain/ws/s/WSSpo.java
@@ -13,9 +13,9 @@ public class WSSpo {
     private String codeCo;
     private Long sCompanyCoId;
     private Date dateOrder;
-    private Long totalAmount;
-    private Long taxRate;
-    private Long exchange;
+    private Float totalAmount;
+    private Float taxRate;
+    private Float exchange;
     private String remark;
     private String freightTerm;
     private String paymentTerm;
@@ -71,22 +71,22 @@ public class WSSpo {
 	public void setDateOrder(Date dateOrder) {
 		this.dateOrder = dateOrder;
 	}
-	public Long getTotalAmount() {
+	public Float getTotalAmount() {
 		return totalAmount;
 	}
-	public void setTotalAmount(Long totalAmount) {
+	public void setTotalAmount(Float totalAmount) {
 		this.totalAmount = totalAmount;
 	}
-	public Long getTaxRate() {
+	public Float getTaxRate() {
 		return taxRate;
 	}
-	public void setTaxRate(Long taxRate) {
+	public void setTaxRate(Float taxRate) {
 		this.taxRate = taxRate;
 	}
-	public Long getExchange() {
+	public Float getExchange() {
 		return exchange;
 	}
-	public void setExchange(Long exchange) {
+	public void setExchange(Float exchange) {
 		this.exchange = exchange;
 	}
 	public String getRemark() {
